+++ pkg/utils/util_darwin.go
@@ -17,7 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package kubelet
+package utils
 
 import "encoding/binary"
